chore(core): move logic in package
