+++ packages/core/.prettierignore
@@ -1,0 +1,2 @@
+dist
+CHANGELOG.md
