+++ packages/core/CHANGELOG.md
@@ -1,0 +1,62 @@
+# Change Log
+
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [0.2.1](https://github.com/balles/diagrams-ts/compare/v0.2.0...v0.2.1) (2021-04-20)
+
+**Note:** Version bump only for package diagrams-ts
+
+
+
+
+
+# [0.2.0](https://github.com/balles/diagrams-ts/compare/v0.1.0...v0.2.0) (2021-04-20)
+
+
+### Features
+
+* **plugin:** move image retrieval to a plugin ([#31](https://github.com/balles/diagrams-ts/issues/31)) ([05adec1](https://github.com/balles/diagrams-ts/commit/05adec1229af73097b483c2756543834f9516d8c))
+
+
+
+
+
+# [0.1.0](https://github.com/balles/diagrams-ts/compare/v0.0.5...v0.1.0) (2021-04-18)
+
+
+### Features
+
+* **renderer:** add wasm renderer ([#30](https://github.com/balles/diagrams-ts/issues/30)) ([4eb9472](https://github.com/balles/diagrams-ts/commit/4eb9472b5c6b206ee6ddc455f0748fc4fb375f82))
+
+
+
+
+
+## 0.0.5 (2021-04-17)
+
+**Note:** Version bump only for package diagrams-ts
+
+
+
+
+
+## 0.0.4 (2021-04-16)
+
+**Note:** Version bump only for package diagrams-ts
+
+
+
+
+
+## 0.0.3 (2021-04-13)
+
+**Note:** Version bump only for package diagrams-ts
+
+
+
+
+
+## 0.0.2 (2021-04-13)
+
+**Note:** Version bump only for package diagrams-ts
