+++ packages/core/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "diagrams-ts",
+  "name": "@diagrams-ts/core",
   "version": "0.2.1",
   "description": "TypeScript port of diagrams",
   "main": "dist/diagrams.js",
@@ -10,13 +10,15 @@
   "diagrams-python-sha": "5489fec1fa3aa5cc54d145860b6b40488106fbfd",
   "repository": {
     "type": "git",
-    "url": "https://github.com/balles/diagrams-ts.git"
+    "url": "https://github.com/balles/diagrams-ts.git",
+    "directory": "packages/core"
   },
   "scripts": {
     "build": "yarn tsc",
     "lint": "yarn prettier . --check && yarn eslint . --ext .ts",
-    "test": "echo 'no test specified'",
-    "format": "yarn prettier . --write"
+    "test": "yarn jest ./src",
+    "format": "yarn prettier . --write",
+    "generate-providers": "yarn ts-node scripts/generate-providers master && yarn format"
   },
   "devDependencies": {
     "@types/jest": "26.0.22",
@@ -27,14 +29,11 @@
     "@typescript-eslint/parser": "4.6.0",
     "eslint": "7.12.1",
     "jest": "26.6.3",
+    "node-fetch": "2.6.1",
     "prettier": "2.1.2",
+    "rimraf": "3.0.2",
     "ts-jest": "26.5.4",
     "ts-node": "9.0.0",
     "typescript": "4.1.2"
-  },
-  "dependencies": {
-    "@diagrams-ts/core": "^0.2.0",
-    "@diagrams-ts/graphviz-cli-renderer": "^0.2.0",
-    "@diagrams-ts/graphviz-functional-ts": "^0.2.0"
   }
 }
