+++ packages/core/scripts/generate-providers.ts
