+++ packages/core/scripts/generation-config.json
