+++ packages/core/src/diagrams.spec.ts
