+++ packages/core/tsconfig.json
@@ -10,11 +10,12 @@
     "declaration": true,
     "resolveJsonModule": true
   },
+  "include": ["src/**/*", "packages/graphviz-ts/graph"],
   "exclude": [
     "node_modules",
     "dist",
     "**/*.spec.ts",
     "examples/**/*",
-    "packages/core/scripts/**/*"
+    "packages/diagrams-ts/scripts/**/*"
   ]
 }
